+++ CHANGELOG.md
@@ -73,6 +73,7 @@ Decidim::Blogs::Post.find_each(&:add_to_index_as_search_resource)
 - **decidim-core**: Fix: Search box on mobile (menu) [#5502](https://github.com/decidim/decidim/pull/5502)
 - **decidim-core**: Fix dynamic controller extensions (undefined method `current_user`) [#5533](https://github.com/decidim/decidim/pull/5533)
 - **decidim-proposals**: Standardize proposal answer callout styles [#5530](https://github.com/decidim/decidim/pull/5530)
+- **decidim-proposals**: Enforce singular group endorsements even with different users [#5579](https://github.com/decidim/decidim/pull/5579)
 
 **Removed**: